Revert "feat(codegen): router rename api"
