+++ .gitignore
@@ -35,9 +35,9 @@ cmd/generate/_example/internal/apps/demoapp/*
 # 但不忽略 config 目录
 !cmd/generate/_example/internal/apps/demoapp/config/
 
-# 不忽略 api/enter.go（注意先保 api 再保文件）
-!cmd/generate/_example/internal/apps/demoapp/api/
-!cmd/generate/_example/internal/apps/demoapp/api/enter.go
+# 不忽略 router/enter.go（注意先保 router 再保文件）
+!cmd/generate/_example/internal/apps/demoapp/router/
+!cmd/generate/_example/internal/apps/demoapp/router/enter.go
 
 # 不忽略 code/enter.go（注意先保 code 再保文件）
 !cmd/generate/_example/internal/apps/demoapp/code/
