+++ cmd/generate/_example/internal/apps/demoapp/router/enter.go
@@ -1,4 +1,4 @@
-package api
+package router
 
 import "github.com/gin-gonic/gin"
