+++ cmd/generate/template/module/router.go.tpl
@@ -1,4 +1,4 @@
-package api
+package router
 
 import (
 	"{{.AppPathInProject}}/controller/ctr{{.PackageName}}"
