+++ go.mod
@@ -4,23 +4,22 @@ go 1.23.3
 
 require (
 	github.com/morehao/go-gitignore v0.0.0-20240726143401-bffa9ffa53e3
-	github.com/morehao/golib v1.2.12
-	github.com/spf13/cobra v1.9.1
-	golang.org/x/mod v0.24.0
+	github.com/morehao/golib v1.2.11
+	github.com/spf13/cobra v1.8.1
+	golang.org/x/mod v0.17.0
 	gorm.io/driver/mysql v1.5.7
-	gorm.io/gorm v1.30.0
+	gorm.io/gorm v1.26.1
 )
 
 require (
-	filippo.io/edwards25519 v1.1.0 // indirect
-	github.com/go-sql-driver/mysql v1.9.2 // indirect
+	github.com/go-sql-driver/mysql v1.7.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
-	github.com/spf13/pflag v1.0.6 // indirect
-	golang.org/x/text v0.25.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/text v0.22.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
